Move all js files except the main router.js entry point to the lib/ folder
