+++ lib/animated-routing-mixin.js
@@ -1,5 +1,5 @@
-import RouteTreeNode from './lib/route-tree-node.js';
-import {Context} from './lib/page.js';
+import RouteTreeNode from './route-tree-node.js';
+import {Context} from './page.js';
 import basicRoutingMixin from './routing-mixin.js';
 import BasicRoutingInterface from './routing-interface.js';
