+++ lib/route-tree-node.js
@@ -20,7 +20,7 @@
 
 import {Context} from './page.js';
 import RouteData from './route-data.js';
-import BasicRoutingInterface from '../routing-interface.js';
+import BasicRoutingInterface from './routing-interface.js';
 
 class RouteTreeNode {
   /** @param {!RouteData} data */
