+++ lib/routing-interface.js
@@ -1,5 +1,5 @@
-import RouteTreeNode from './lib/route-tree-node.js';
-import {Context} from './lib/page.js';
+import RouteTreeNode from './route-tree-node.js';
+import {Context} from './page.js';
 
 /** @interface */
 class BasicRoutingInterface {
