+++ lib/routing-mixin.js
@@ -1,6 +1,6 @@
-import RouteTreeNode from './lib/route-tree-node.js';
-import RouteData from './lib/route-data.js';
-import {Context} from './lib/page.js';
+import RouteTreeNode from './route-tree-node.js';
+import RouteData from './route-data.js';
+import {Context} from './page.js';
 import BasicRoutingInterface from './routing-interface.js';
 
 /**
