+++ router.js
@@ -17,9 +17,9 @@
 
 import {Context, Page} from './lib/page.js';
 import RouteTreeNode from './lib/route-tree-node.js';
-import routingMixin from './routing-mixin.js';
-import animatedRoutingMixin from './animated-routing-mixin.js';
-import BasicRoutingInterface from './routing-interface.js';
+import routingMixin from './lib/routing-mixin.js';
+import animatedRoutingMixin from './lib/animated-routing-mixin.js';
+import BasicRoutingInterface from './lib/routing-interface.js';
 import RouteData from './lib/route-data.js';
 
 class Router {
